Rename to "rest", and optimize a bit
